Seperate proprietary and ocms js asset bundles (#4075)
* register a new bundle

* add the new bundle target asset

* remove the non-proprietary source file from the proprietary bundle definition

* create a new build target for the non-proprietary asset files

* remove the non-proprietary code from the compiled proprietary asset

* add compiled and minified build-ocms file

* add missing line at the end

* use better asset file names
