+++ modules/backend/ServiceProvider.php
@@ -75,6 +75,7 @@ class ServiceProvider extends ModuleServiceProvider
             $combiner->registerBundle('~/modules/backend/formwidgets/codeeditor/assets/js/build.js');
             $combiner->registerBundle('~/modules/backend/formwidgets/fileupload/assets/less/fileupload.less');
             $combiner->registerBundle('~/modules/backend/formwidgets/nestedform/assets/less/nestedform.less');
+            $combiner->registerBundle('~/modules/backend/formwidgets/richeditor/assets/js/build-oc-plugins.js');
 
             /*
              * Rich Editor is protected by DRM
