+++ modules/backend/formwidgets/richeditor/assets/js/build-oc-plugins.js
@@ -1,0 +1,15 @@
+/*
+ * This is a bundle file, you can compile this in two ways:
+ * (1) Using your favorite JS combiner
+ * (2) Using CLI command:
+ *   php artisan october:util compile assets
+ *
+ * @see build-oc-plugins-min.js
+ *
+
+=require plugins/mediamanager.js
+=require plugins/pagelinks.js
+=require plugins/figures.js
+=require richeditor.js
+
+*/
