+++ modules/backend/formwidgets/richeditor/assets/js/build.js
@@ -42,9 +42,4 @@
 // More testing needed
 //require ../vendor/froala_drm/js/plugins/quick_insert.js
 
-=require plugins/mediamanager.js
-=require plugins/pagelinks.js
-=require plugins/figures.js
-=require richeditor.js
-
 */
